Compiled without warnings on 32 and 64 bit Windows
git-svn-id: https://groke.otago.ac.nz/ant/ant@172 7489a7b4-4588-4892-8acd-774bacfab51c

+++ Link-The-Wiki/link_parts.h
@@ -70,12 +70,12 @@ return file;
 	STRNNEW()
 	---------
 */
-inline static char *strnnew(char *source, long length)
+inline static char *strnnew(char *source, long long length)
 {
 char *ans;
 
-ans = new char [length + 1];
-strncpy(ans, source, length);
+ans = new char [(size_t)(length + 1)];
+strncpy(ans, source, (size_t)length);
 ans[length] = '\0';
 
 return ans;
